backend/requirements.txt
# AI Code Reviewer - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Models
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# AI/ML Dependencies
transformers==4.35.2
torch==2.1.1
langchain==0.1.0
langchain-community==0.0.10

# Utilities
python-dateutil==2.8.2